case-sensitive and case-insensitive collisions. THIS WAS SUCH A FUCKING MESS. had to use the browser version fo the gihub to delete this file and to commit it.
warning: the following paths have collided (e.g. case-sensitive paths
on a case-insensitive filesystem) and only one from the same
colliding group is in the working tree


THIS WAS SUCH A FUCKING MESS. had to use the browser version fo the gihub to delete this file and to commit it.